+++ src/App.ts
@@ -10,11 +10,12 @@ import Context from "./App/Context";
 export interface Class extends Document.Class {
 	name:string;
 	description?:string;
-	allowsOrigins:(Pointer | string)[];
+	allowsOrigins?:(Pointer | string)[];
 }
 
 export const RDF_CLASS:string = NS.CS.Class.Application;
 
+// TODO Problem in circular reference: the SCHEMA is hardcoded in SDKContext
 export const SCHEMA:ObjectSchema.Class = {
 	"name": {
 		"@id": NS.CS.Predicate.name,