+++ src/LDP.ts
@@ -1,3 +1,4 @@
+import * as AddMemberAction from "./LDP/AddMemberAction";
 import * as BasicContainer from "./LDP/BasicContainer";
 import * as Container from "./LDP/Container";
 import * as DirectContainer from "./LDP/DirectContainer";
@@ -6,6 +7,7 @@ import * as PersistedContainer from "./LDP/PersistedContainer";
 import * as RDFSource from "./LDP/RDFSource";
 
 export {
+	AddMemberAction,
 	BasicContainer,
 	Container,
 	DirectContainer,