+++ src/LDP/PersistedContainer.ts
@@ -5,16 +5,32 @@ import * as Pointer from "./../Pointer";
 import * as Utils from "./../Utils";
 
 export interface Class extends PersistedDocument.Class {
+	addMember( member:Pointer.Class ): Promise<HTTP.Response.Class>;
+	addMember( memberURI:string ): Promise<HTTP.Response.Class>;
+
+	addMembers( members:(Pointer.Class | string)[] ): Promise<HTTP.Response.Class>;
+
 	createChild( slug:string, object:Object ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 	createChild( slug:string ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 	createChild( object:Object ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 	createChild():Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 
+	getMembers( includeNonReadable?:boolean ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+
 	upload( slug:string, blob:Blob ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 	upload( blob:Blob ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
+}
 
+function addMember( member:Pointer.Class ): Promise<HTTP.Response.Class>;
+function addMember( memberURI:string ): Promise<HTTP.Response.Class>;
+function addMember( memberOrUri:any ): Promise<HTTP.Response.Class> {
+	let that:PersistedDocument.Class = <PersistedDocument.Class> this;
+	return that._documents.addMember( that.id, memberOrUri );
+}
 
-	getMembers( includeNonReadable?:boolean ):Promise<[ Pointer.Class[], HTTP.Response.Class ]>;
+function addMembers( members:(Pointer.Class | string)[] ): Promise<HTTP.Response.Class> {
+	let that:PersistedDocument.Class = <PersistedDocument.Class> this;
+	return that._documents.addMembers( that.id, members );
 }
 
 function createChild( slug:string, object:Object ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
@@ -33,6 +49,10 @@ function createChild( slugOrObject?:any, object?:Object ):Promise<[ Pointer.Clas
 	}
 }
 
+function getMembers( includeNonReadable:boolean = true ):Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
+	return this._documents.getMembers( this.id, includeNonReadable );
+}
+
 function upload( slug:string, blob:Blob ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 function upload( blob:Blob ):Promise<[ Pointer.Class, HTTP.Response.Class ]>;
 function upload( slugOrBlob:any, blob:any = null ):Promise<[ Pointer.Class, HTTP.Response.Class ]> {
@@ -46,13 +66,11 @@ function upload( slugOrBlob:any, blob:any = null ):Promise<[ Pointer.Class, HTTP
 	}
 }
 
-function getMembers( includeNonReadable:boolean = true ):Promise<[ Pointer.Class[], HTTP.Response.Class ]> {
-	return this._documents.getMembers( this.id, includeNonReadable );
-}
-
 	export class Factory {
 	static hasClassProperties( document:Document.Class ):boolean {
 		return Utils.hasFunction( document, "createChild" )
+			&& Utils.hasFunction( document, "addMember" )
+			&& Utils.hasFunction( document, "addMembers" )
 			&& Utils.hasFunction( document, "upload" )
 			&& Utils.hasFunction( document, "getMembers" );
 	}
@@ -61,17 +79,23 @@ function getMembers( includeNonReadable:boolean = true ):Promise<[ Pointer.Class
 		if( Factory.hasClassProperties( persistedDocument ) ) return <any> persistedDocument;
 
 		Object.defineProperties( persistedDocument, {
+			"addMember": {
+				writable: false,
+				enumerable: false,
+				configurable: true,
+				value: addMember,
+			},
+			"addMembers": {
+				writable: false,
+				enumerable: false,
+				configurable: true,
+				value: addMembers,
+			},
 			"createChild": {
 				writable: false,
 				enumerable: false,
 				configurable: true,
 				value: createChild,
-			},
-			"upload": {
-				writable: false,
-				enumerable: false,
-				configurable: true,
-				value: upload,
 			},
 			"getMembers": {
 				writable: false,
@@ -79,8 +103,16 @@ function getMembers( includeNonReadable:boolean = true ):Promise<[ Pointer.Class
 				configurable: true,
 				value: getMembers,
 			},
+			"upload": {
+				writable: false,
+				enumerable: false,
+				configurable: true,
+				value: upload,
+			},
 		} );
 
 		return <any> persistedDocument;
 	}
 }
+
+export default Class;