+++ src/NS/C.spec.ts
@@ -39,7 +39,7 @@ describe( module(
 		it( isDefined(), ():void => {
 			expect( C.Class ).toBeDefined();
 			expect( Utils.isFunction( C.Class ) ).toBe( true );
-			expect( Object.keys( C.Class ).length ).toBe( 9 );
+			expect( Object.keys( C.Class ).length ).toBe( 10 );
 		});
 
 		it( hasProperty(
@@ -141,6 +141,17 @@ describe( module(
 			expect( C.Class.RDFRepresentation ).toBe( "https://carbonldp.com/ns/v1/platform#RDFRepresentation" );
 		});
 
+		it( hasProperty(
+			STATIC,
+			"AddMemberAction",
+			"string"
+		), ():void => {
+			expect( C.Class.AddMemberAction ).toBeDefined();
+			expect( Utils.isString( C.Class.AddMemberAction ) ).toBe( true );
+
+			expect( C.Class.AddMemberAction ).toBe( "https://carbonldp.com/ns/v1/platform#AddMemberAction" );
+		});
+
 	});
 
 	describe( clazz(
@@ -152,7 +163,7 @@ describe( module(
 			expect( C.Predicate ).toBeDefined();
 			expect( Utils.isFunction( C.Predicate ) ).toBe( true );
 
-			expect( Object.keys( C.Predicate ).length ).toBe( 7 );
+			expect( Object.keys( C.Predicate ).length ).toBe( 8 );
 		});
 
 		it( hasProperty(
@@ -232,6 +243,17 @@ describe( module(
 			expect( Utils.isString( C.Predicate.size ) ).toBe( true );
 
 			expect( C.Predicate.size ).toBe( "https://carbonldp.com/ns/v1/platform#size" );
+		});
+
+		it( hasProperty(
+			STATIC,
+			"targetMember",
+			"string"
+		), ():void => {
+			expect( C.Predicate.targetMember ).toBeDefined();
+			expect( Utils.isString( C.Predicate.targetMember ) ).toBe( true );
+
+			expect( C.Predicate.targetMember ).toBe( "https://carbonldp.com/ns/v1/platform#targetMember" );
 		});
 
 	});