+++ src/PersistedApp.ts
@@ -9,7 +9,7 @@ export interface Class extends PersistedDocument.Class {
 	name:string;
 	description?:string;
 	rootContainer:LDP.PersistedContainer.Class;
-	allowsOrigins:(Pointer | string)[];
+	allowsOrigins?:(Pointer | string)[];
 }
 
 export class Factory {