+++ src/SDKContext.ts
@@ -1,3 +1,4 @@
+import * as AddMemberAction from "./LDP/AddMemberAction";
 import * as App from "./App";
 import * as APIDescription from "./APIDescription";
 import * as Auth from "./Auth";
@@ -168,11 +169,13 @@ export class Class implements Context {
 				"@id": NS.CS.Predicate.rootContainer,
 				"@type": "@id",
 			},
-			"allowsOrigin": {
+			"allowsOrigins": {
 				"@id": NS.CS.Predicate.allowsOrigin,
+				"@container": "@set",
 			},
 		});
 
+		this.extendObjectSchema( AddMemberAction.RDF_CLASS, AddMemberAction.SCHEMA );
 		this.extendObjectSchema( Auth.Token.RDF_CLASS, Auth.Token.CONTEXT );
 
 		this.extendObjectSchema( Agent.RDF_CLASS, Agent.SCHEMA );