+++ CHANGELOG.md
@@ -1,3 +1,12 @@
+# 0.27.0 (April 12, 2016)
+- Added `context.documents.addMember()` method, which lets you add a single member reference to a Container
+- Added `context.documents.addMember()` method, same as `persistedContainer.addMember()`
+- Added `context.documents.addMembers()` method, which lets you add a multiple members references to a Container
+- Added `context.documents.addMembers()` method, same as `persistedContainer.addMembers()`
+
+#### Breaking Changes
+- Changed `context.documents.delete()` method, now it accepts a document URI and not a PersistedDocument object
+
 # 0.26.0 (April 12, 2016)
 - Added `persistedContainer.getMembers()` method, which lets you obtain the members of the Container